+++ cedar-policy-core/src/ast/policy.rs
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-use crate::ast::*;
+use crate::parser::err::ParseErrors;
+use crate::{ast::*, FromNormalizedStr};
 use itertools::Itertools;
 use miette::Diagnostic;
 use serde::{Deserialize, Serialize};
@@ -63,7 +64,7 @@ impl Template {
     /// Construct a `Template` from its components
     pub fn new(
         id: PolicyID,
-        annotations: BTreeMap<Id, SmolStr>,
+        annotations: BTreeMap<AnyId, SmolStr>,
         effect: Effect,
         principal_constraint: PrincipalConstraint,
         action_constraint: ActionConstraint,
@@ -123,12 +124,12 @@ impl Template {
     }
 
     /// Get data from an annotation.
-    pub fn annotation(&self, key: &Id) -> Option<&SmolStr> {
+    pub fn annotation(&self, key: &AnyId) -> Option<&SmolStr> {
         self.body.annotation(key)
     }
 
     /// Get all annotation data.
-    pub fn annotations(&self) -> impl Iterator<Item = (&Id, &SmolStr)> {
+    pub fn annotations(&self) -> impl Iterator<Item = (&AnyId, &SmolStr)> {
         self.body.annotations()
     }
 
@@ -366,12 +367,12 @@ impl Policy {
     }
 
     /// Get data from an annotation.
-    pub fn annotation(&self, key: &Id) -> Option<&SmolStr> {
+    pub fn annotation(&self, key: &AnyId) -> Option<&SmolStr> {
         self.template.annotation(key)
     }
 
     /// Get all annotation data.
-    pub fn annotations(&self) -> impl Iterator<Item = (&Id, &SmolStr)> {
+    pub fn annotations(&self) -> impl Iterator<Item = (&AnyId, &SmolStr)> {
         self.template.annotations()
     }
 
@@ -672,12 +673,12 @@ impl StaticPolicy {
     }
 
     /// Get data from an annotation.
-    pub fn annotation(&self, key: &Id) -> Option<&SmolStr> {
+    pub fn annotation(&self, key: &AnyId) -> Option<&SmolStr> {
         self.0.annotation(key)
     }
 
     /// Get all annotation data.
-    pub fn annotations(&self) -> impl Iterator<Item = (&Id, &SmolStr)> {
+    pub fn annotations(&self) -> impl Iterator<Item = (&AnyId, &SmolStr)> {
         self.0.annotations()
     }
 
@@ -734,10 +735,10 @@ impl StaticPolicy {
         self.0.condition()
     }
 
-    /// Construct a `Policy` from its components
+    /// Construct a `StaticPolicy` from its components
     pub fn new(
         id: PolicyID,
-        annotations: BTreeMap<Id, SmolStr>,
+        annotations: BTreeMap<AnyId, SmolStr>,
         effect: Effect,
         principal_constraint: PrincipalConstraint,
         action_constraint: ActionConstraint,
@@ -789,16 +790,106 @@ impl From<StaticPolicy> for Arc<Template> {
     }
 }
 
-/// Policy datatype.
-///
-/// This will also represent the serialized form of policies on disk
-/// and on the network.
+/// Like `Id`, except this specifically _can_ contain Cedar reserved identifiers.
+/// (It still can't contain, for instance, spaces or characters like '+'.)
+//
+// For now, internally, `AnyId`s are just owned `SmolString`s.
+#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone, Hash, PartialOrd, Ord)]
+pub struct AnyId(SmolStr);
+
+impl AnyId {
+    /// Create a new `AnyId` from a `String`, where it is the caller's
+    /// responsibility to ensure that the string is indeed a valid `AnyId`.
+    ///
+    /// When possible, callers should not use this, and instead use `s.parse()`,
+    /// which checks that `s` is a valid `AnyId`, and returns a parse error
+    /// if not.
+    ///
+    /// This method was created for the `From<cst::Ident> for AnyId` impl to use.
+    /// See notes on `Id::new_unchecked()`.
+    pub(crate) fn new_unchecked(s: impl Into<SmolStr>) -> AnyId {
+        AnyId(s.into())
+    }
+
+    /// Get the underlying string
+    pub fn into_smolstr(self) -> SmolStr {
+        self.0
+    }
+}
+
+impl AsRef<str> for AnyId {
+    fn as_ref(&self) -> &str {
+        &self.0
+    }
+}
+
+impl std::fmt::Display for AnyId {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", &self.0)
+    }
+}
+
+// allow `.parse()` on a string to make an `AnyId`
+impl std::str::FromStr for AnyId {
+    type Err = ParseErrors;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        crate::parser::parse_anyid(s)
+    }
+}
+
+impl FromNormalizedStr for AnyId {
+    fn describe_self() -> &'static str {
+        "AnyId"
+    }
+}
+
+#[cfg(feature = "arbitrary")]
+impl<'a> arbitrary::Arbitrary<'a> for AnyId {
+    fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
+        // AnyId syntax:
+        // ['_''a'-'z''A'-'Z']['_''a'-'z''A'-'Z''0'-'9']*
+
+        let construct_list = |s: &str| s.chars().collect::<Vec<char>>();
+        let list_concat = |s1: &[char], s2: &[char]| [s1, s2].concat();
+        // the set of the first character of an AnyId
+        let head_letters = construct_list("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz_");
+        // the set of the remaining characters of an AnyId
+        let tail_letters = list_concat(&construct_list("0123456789"), &head_letters);
+        // identifier character count minus 1
+        let remaining_length = u.int_in_range(0..=16)?;
+        let mut cs = vec![*u.choose(&head_letters)?];
+        cs.extend(
+            (0..remaining_length)
+                .map(|_| u.choose(&tail_letters))
+                .collect::<Result<Vec<&char>, _>>()?,
+        );
+        let s: String = cs.into_iter().collect();
+        crate::parser::parse_ident(&s).unwrap_or_else(|e| panic!("all strings constructed this way should be valid AnyIds, but this one is not: {s:?}\nerror: {e}"));
+        Ok(Self::new_unchecked(s))
+    }
+
+    fn size_hint(depth: usize) -> (usize, Option<usize>) {
+        arbitrary::size_hint::and_all(&[
+            // for arbitrary length
+            <usize as arbitrary::Arbitrary>::size_hint(depth),
+            // for arbitrary choices
+            // we use the size hint of a vector of `u8` to get an underestimate of bytes required by the sequence of choices.
+            <Vec<u8> as arbitrary::Arbitrary>::size_hint(depth),
+        ])
+    }
+}
+
+/// Policy datatype. This is used for both templates (in which case it contains
+/// slots) and static policies (in which case it contains zero slots).
 #[derive(Serialize, Deserialize, Clone, Hash, Eq, PartialEq, Debug)]
 pub struct TemplateBody {
     /// ID of this policy
     id: PolicyID,
-    /// Annotations available for external applications, as key-value store
-    annotations: BTreeMap<Id, SmolStr>,
+    /// Annotations available for external applications, as key-value store.
+    /// Note that the keys are `AnyId`, so Cedar reserved words like `if` and `has`
+    /// are explicitly allowed as annotations.
+    annotations: BTreeMap<AnyId, SmolStr>,
     /// `Effect` of this policy
     effect: Effect,
     /// Head constraint for principal. This will be a boolean-valued expression:
@@ -839,12 +930,12 @@ impl TemplateBody {
     }
 
     /// Get data from an annotation.
-    pub fn annotation(&self, key: &Id) -> Option<&SmolStr> {
+    pub fn annotation(&self, key: &AnyId) -> Option<&SmolStr> {
         self.annotations.get(key)
     }
 
     /// Get all annotation data.
-    pub fn annotations(&self) -> impl Iterator<Item = (&Id, &SmolStr)> {
+    pub fn annotations(&self) -> impl Iterator<Item = (&AnyId, &SmolStr)> {
         self.annotations.iter()
     }
 
@@ -913,7 +1004,7 @@ impl TemplateBody {
     /// Construct a `Policy` from its components
     pub fn new(
         id: PolicyID,
-        annotations: BTreeMap<Id, SmolStr>,
+        annotations: BTreeMap<AnyId, SmolStr>,
         effect: Effect,
         principal_constraint: PrincipalConstraint,
         action_constraint: ActionConstraint,